MULTI-AGENT-CANDIDATE-SELECTION/requirements.txt
# Core dependencies for RAG System

# PyTorch - Core ML framework (compatible versions)
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0

# LlamaIndex - Core RAG framework
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-groq>=0.1.0
llama-index-llms-gemini>=0.1.0
llama-index-vector-stores-chroma>=0.1.0

# Vector Database
chromadb>=0.4.18

# PDF processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pymupdf>=1.23.0

# NLP and ML (compatible versions)
sentence-transformers>=2.3.0,<3.0.0
transformers>=4.35.0,<5.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0,<2.0.0

# Web interface
streamlit>=1.29.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
tqdm>=4.66.0

# Groq SDK (for LLM)
groq>=0.4.0

# Google Generative AI (for Gemini fallback)
google-generativeai>=0.3.0
